src/requirements.txt
asyncpg==0.17.0
Django==2.0.6
fuzzywuzzy==0.16.0
psycopg2==2.7.5
python-Levenshtein==0.12.0
PyYAML==3.13
git+git://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
